third_party/blink/web_tests/platform/generic/fast/multicol/event-offset-complex-tree-expected.txt

Dispatch mouse click events in each column and check the result.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.target.id is event.expectedElementID
PASS event.offsetX is event.expectedOffsetX
PASS event.offsetY is event.expectedOffsetY
PASS event.target.id is event.expectedElementID
PASS event.offsetX is event.expectedOffsetX
PASS event.offsetY is event.expectedOffsetY
PASS event.target.id is event.expectedElementID
PASS event.offsetX is event.expectedOffsetX
PASS event.offsetY is event.expectedOffsetY





first


second



third
